backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Vector Database & Embeddings
qdrant-client==1.7.0
sentence-transformers==2.2.2
numpy==1.24.3

# Text Processing
langchain==0.0.350
langchain-community==0.0.4
tiktoken==0.5.2
pypdf==3.17.1
python-docx==1.1.0
markdown==3.5.1

# Storage
minio==7.2.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0
python-json-logger==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0